+++ packages/jaspr/lib/src/framework/render_object.dart
@@ -8,6 +8,8 @@ part of 'framework.dart';
 typedef RenderObjectVisitor = void Function(RenderObject child);
 
 abstract class RenderObject {
+  RenderObject? get parent;
+
   RenderObject createChildRenderObject();
 
   void updateElement(String tag, String? id, String? classes, Map<String, String>? styles,
@@ -17,49 +19,34 @@ abstract class RenderObject {
 
   void skipChildren();
 
-  void attach(covariant RenderObject? parent, covariant RenderObject? after);
+  void attach(covariant RenderObject child, {covariant RenderObject? after});
 
-  void remove();
+  void remove(covariant RenderObject child);
 }
 
-abstract class MultiChildRenderObjectElement extends MultiChildElement with RenderObjectElement {
-  MultiChildRenderObjectElement(super.component);
-
-  @override
-  void update(covariant Component newComponent) {
-    super.update(newComponent);
-    _dirty = true;
-    rebuild();
-  }
-}
-
-abstract class SingleChildRenderObjectElement extends SingleChildElement with RenderObjectElement {
-  SingleChildRenderObjectElement(super.component);
-
-  @override
-  void update(covariant Component newComponent) {
-    super.update(newComponent);
-    _dirty = true;
-    rebuild();
-  }
-}
+abstract class BuildableRenderObjectElement = BuildableElement with RenderObjectElement;
+abstract class ProxyRenderObjectElement = ProxyElement with RenderObjectElement;
+abstract class LeafRenderObjectElement = LeafElement with RenderObjectElement;
 
 mixin RenderObjectElement on Element {
+  RenderObject createRenderObject() {
+    var renderObject = _parentRenderObjectElement!.renderObject.createChildRenderObject();
+    assert(renderObject.parent == _parentRenderObjectElement!.renderObject);
+    return renderObject;
+  }
+
   void updateRenderObject();
 
   RenderObject get renderObject => _renderObject!;
   RenderObject? _renderObject;
 
   @override
-  void _firstBuild([VoidCallback? onBuilt]) {
+  void didMount() {
     if (_renderObject == null) {
-      _renderObject = _parentRenderObjectElement!.renderObject.createChildRenderObject();
+      _renderObject = createRenderObject();
       updateRenderObject();
     }
-    super._firstBuild(() {
-      attachRenderObject();
-      onBuilt?.call();
-    });
+    super.didMount();
   }
 
   @override
@@ -68,17 +55,26 @@ mixin RenderObjectElement on Element {
     updateRenderObject();
   }
 
+  @override
   void attachRenderObject() {
-    Element? prevElem = _prevAncestorSibling;
-    while (prevElem != null && prevElem._lastRenderObjectElement == null) {
-      prevElem = prevElem._prevAncestorSibling;
+    var parent = _parentRenderObjectElement?.renderObject;
+    if (parent != null) {
+      Element? prevElem = _prevAncestorSibling;
+      while (prevElem != null && prevElem._lastRenderObjectElement == null) {
+        prevElem = prevElem._prevAncestorSibling;
+      }
+      var after = prevElem?._lastRenderObjectElement;
+      parent.attach(renderObject, after: after?.renderObject);
+      assert(renderObject.parent == parent);
     }
-    var after = prevElem?._lastRenderObjectElement;
-    renderObject.attach(_parentRenderObjectElement?.renderObject, after?.renderObject);
   }
 
   void detachRenderObject() {
-    renderObject.remove();
+    var parent = _parentRenderObjectElement?.renderObject;
+    if (parent != null) {
+      parent.remove(renderObject);
+      assert(renderObject.parent == null);
+    }
   }
 
   @override