+++ packages/jaspr/lib/src/framework/stateful_component.dart
@@ -583,7 +583,7 @@ mixin PreloadStateMixin<T extends StatefulComponent> on State<T> {
 }
 
 /// An [Element] that uses a [StatefulComponent] as its configuration.
-class StatefulElement extends MultiChildElement {
+class StatefulElement extends BuildableElement {
   /// Creates an element that uses the given component as its configuration.
   StatefulElement(StatefulComponent component)
       : _state = component.createState(),
@@ -620,8 +620,10 @@ class StatefulElement extends MultiChildElement {
   State? _state;
   State get state => _state!;
 
+  Future? _asyncInitState;
+
   @override
-  void _firstBuild([VoidCallback? onBuilt]) {
+  void didMount() {
     assert(state._debugLifecycleState == _StateLifecycle.created);
 
     // We check if state uses on of the mixins that support async initialization,
@@ -629,20 +631,13 @@ class StatefulElement extends MultiChildElement {
     // In this case we don't call [_initState()] directly here, but rather let it
     // be called by the mixins implementation.
 
-    Future? asyncFirstBuild;
-
     if (owner.isFirstBuild && state is PreloadStateMixin && !binding.isClient) {
-      asyncFirstBuild = (state as PreloadStateMixin).preloadState();
-    }
-
-    if (asyncFirstBuild != null) {
-      _asyncFirstBuild = asyncFirstBuild.then((_) => _initState());
-      asyncFirstBuild.whenComplete(() => _asyncFirstBuild = null);
+      _asyncInitState = (state as PreloadStateMixin).preloadState().then((_) => _initState());
     } else {
       _initState();
     }
 
-    super._firstBuild(onBuilt);
+    super.didMount();
   }
 
   void _initState() {
@@ -655,7 +650,7 @@ class StatefulElement extends MultiChildElement {
         '${state.runtimeType}.initState() returned a Future.\n\n'
         'Rather than awaiting on asynchronous work directly inside of initState, '
         'call a separate method to do this work without awaiting it.\n\n'
-        'If you need to do some async work before the first render, use PreloadStateMixin or DeferRenderMixin on State.',
+        'If you need to do some async work before the first render, use PreloadStateMixin on State.',
       );
     } finally {
       _debugSetAllowIgnoredCallsToMarkNeedsBuild(false);
@@ -672,25 +667,33 @@ class StatefulElement extends MultiChildElement {
   }
 
   @override
-  void performRebuild() {
+  Object? performRebuild() {
+    if (owner.isFirstBuild && _asyncInitState != null) {
+      return _asyncInitState!.then((_) {
+        if (_didChangeDependencies) {
+          state.didChangeDependencies();
+          _didChangeDependencies = false;
+        }
+        super.performRebuild();
+      });
+    }
     if (_didChangeDependencies) {
       state.didChangeDependencies();
       _didChangeDependencies = false;
     }
     super.performRebuild();
+    return null;
   }
 
   @override
   void update(StatefulComponent newComponent) {
     super.update(newComponent);
     assert(component == newComponent);
-    final StatefulComponent oldComponent = state.component;
-
-    // We mark ourselves as dirty before calling didUpdateWidget to
-    // let authors call setState from within didUpdateWidget without triggering
-    // asserts.
-    _dirty = true;
     state._component = newComponent;
+  }
+
+  @override
+  void didUpdate(StatefulComponent oldComponent) {
     try {
       _debugSetAllowIgnoredCallsToMarkNeedsBuild(true);
       // TODO: check for returned future
@@ -698,7 +701,7 @@ class StatefulElement extends MultiChildElement {
     } finally {
       _debugSetAllowIgnoredCallsToMarkNeedsBuild(false);
     }
-    rebuild();
+    super.didUpdate(oldComponent);
   }
 
   @override