+++ packages/jaspr/lib/src/server/adapters/client_component_adapter.dart
@@ -9,12 +9,13 @@ class ClientComponentAdapter extends ElementBoundaryAdapter {
   ClientComponentAdapter(this.name, this.data, super.element);
 
   final String name;
-  final String data;
+  final String? data;
 
   @override
-  void processBoundary(ChildListRange range) {
-    range.start.insertNext(ChildNodeData(MarkupRenderObject()..updateText('<!-- \$$name $data  -->', true)));
-    range.end.insertPrev(ChildNodeData(MarkupRenderObject()..updateText('<!-- /\$$name -->', true)));
+  void applyBoundary(ChildListRange range) {
+    range.start.insertNext(
+        ChildNodeData(MarkupRenderObject()..updateText('<!--\$$name${data != null ? ' data=$data' : ''}-->', true)));
+    range.end.insertPrev(ChildNodeData(MarkupRenderObject()..updateText('<!--/\$$name-->', true)));
   }
 }