+++ packages/jaspr_tailwind/lib/builder.dart
@@ -44,9 +44,9 @@ class TailwindBuilder implements Builder {
         '--input',
         scratchSpace.fileFor(buildStep.inputId).path,
         '--output',
-        scratchSpace.fileFor(outputId).path,
+        scratchSpace.fileFor(outputId).path.toPosix(),
         '--content',
-        p.join(Directory.current.path, '{lib,web}', '**', '*.dart'),
+        p.join(Directory.current.path, '{lib,web}', '**', '*.dart').toPosix(true),
         if (hasCustomConfig) ...[
           '--config',
           p.join(Directory.current.path, 'tailwind.config.js'),
@@ -62,4 +62,14 @@ class TailwindBuilder implements Builder {
   Map<String, List<String>> get buildExtensions => {
         'web/{{file}}.tw.css': ['web/{{file}}.css']
       };
+}
+
+extension POSIXPath on String {
+  String toPosix([bool quoted = false]) {
+    if (Platform.isWindows) {
+      final result = replaceAll('\\', '/');
+      return quoted ? "'$result'" : result;
+    }
+    return this;
+  }
 }